+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+#Version 0.3.0
+
+##Features
+
+- Add Modal demo component with embedded template
+- Add Modal demo component with external template
+
 #Version 0.2.0
 
 ##Features